requirements.txt
pillow 
torch
torchvision 
transformers 
pyiqa